refactor(core): Update UnitType handling and validation
- Added `UnitTypeNameOnlyCommand` to handle unit type creation with only a name.
- Renamed `CreateUnitTypeValidator.cs` to `NameOfUnitTypeValidator.cs`.
  - Removed `ApplyCustomValidation()` method that enforced unique name validation.
- Updated `UnitTypesProfile` mapping:
  - Renamed `CreateUnitTypeMapping.cs` to `CreateUnitTypeMapping.cs`.
  - Replaced `CreateUnitTypeMapping()` with `UnitTypeOnlyMapping()`.
  - Updated profile usage from `CreateUnitTypeMapping();` to `UnitTypeOnlyMapping();`.

These changes streamline UnitType validation and mapping, aligning with the new command structure.
