+++ backend/Core/Featurs/UnitType/Commands/Requests/UnitTypeNameOnlyCommand.cs
@@ -1,0 +1,9 @@
+using Core.Bases;
+using MediatR;
+
+namespace Core.Featurs.UnitType.Commands.Requests;
+
+public class UnitTypeNameOnlyCommand : IRequest<Response<string>>
+{
+    public required string Name { get; set; }
+}
