+++ backend/Core/Mapping/UnitTypes/UnitTypesProfile.cs
@@ -9,6 +9,6 @@ public partial class UnitTypesProfile : Profile
     {
         GetUnitTypesMapping();
         GetUnitTypeMapping();
-        CreateUnitTypeMapping();
+        UnitTypeOnlyMapping();
     }
 }
